CHAD-17087: zwave-fan lazy lading of sub-drivers
