+++ .github/workflows/publish-test-results.yml
@@ -1,0 +1,30 @@
+name: Publish test results
+on:
+  workflow_run:
+    workflows: [Run driver tests]
+    types:
+      - completed
+
+jobs:
+  publish-test-results:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Download artifacts
+        uses: dawidd6/action-download-artifact@v6
+        with:
+          workflow: run-tests.yml
+          run_id: ${{ github.event.workflow_run.id }}
+      - run: echo "pr_number=$(cat pr_number/pr_number.txt)" >> $GITHUB_ENV
+      - name: Publish test results
+        uses: EnricoMi/publish-unit-test-result-action@v2
+        with:
+          commit: ${{ github.event.workflow_run.head_sha }}
+          event_file: event-file/event.json
+          event_name: ${{ github.event.workflow_run.event }}
+          files: "tests/*.xml"
+      - name: Publish coverage results
+        uses: 5monkeys/cobertura-action@master
+        with:
+          pull_request_number: ${{ env.pr_number }}
+          path: "coverage/*.xml"
+          minimum_coverage: 90
