+++ drivers/SmartThings/zwave-fan/src/zwave-fan-4-speed/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FAN_4_SPEED_FINGERPRINTS = {
+  {mfr = 0x001D, prod = 0x0038, model = 0x0002}, -- Leviton 4-Speed Fan Controller
+}
+
+return FAN_4_SPEED_FINGERPRINTS
